fix up formating of messages

+++ files/common/remove.sh
@@ -42,11 +42,28 @@ if [[ ! -z `docker ps -a | grep quorum-dev-quickstart_pet_shop` ]]; then
 fi
 
 docker image rm quorum-dev-quickstart/block-explorer-light:develop
-docker image rm pegasyseng/orion:${QUORUM_ORION_VERSION}
-docker image rm pegasyseng/ethsigner:${QUORUM_ETHSIGNER_VERSION}
-docker image rm hyperledger/besu:${BESU_VERSION}
-docker image rm quorumengineering/quorum:${QUORUM_VERSION}
-docker image rm quorumengineering/tessera:${QUORUM_TESSERA_VERSION}
+docker image rm consensys/quorum-ethsigner:${QUORUM_ETHSIGNER_VERSION}
+if grep -q 'orion:' docker-compose.yml 2> /dev/null ; then
+  docker image rm consensys/quorum-orion:${QUORUM_ORION_VERSION}
+fi
+if grep -q 'besu:' docker-compose.yml 2> /dev/null ; then
+  docker image rm hyperledger/besu:${BESU_VERSION}
+fi
+if grep -q 'tessera:' docker-compose.yml 2> /dev/null ; then
+  docker image rm quorumengineering/tessera:${QUORUM_TESSERA_VERSION}
+fi
+if grep -q 'quorum:' docker-compose.yml 2> /dev/null ; then
+  docker image rm quorumengineering/quorum:${QUORUM_VERSION}
+fi
+if grep -q 'cakeshop:' docker-compose.yml 2> /dev/null ; then
+  docker image rm quorumengineering/cakeshop:${QUORUM_CAKESHOP_VERSION}
+fi
+if grep -q 'kibana:' docker-compose.yml 2> /dev/null ; then
+  docker image rm quorum-test-network_elasticsearch
+  docker image rm quorum-test-network_logstash
+  docker image rm quorum-test-network_filebeat
+  docker image rm quorum-test-network_metricbeat
+fi
 
 rm ${LOCK_FILE}
 echo "Lock file ${LOCK_FILE} removed"
